backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.28.1
anthropic==0.42.0
python-dotenv==1.0.1
firecrawl-py==1.9.0
